Refactor dashboard navigation
[Description]
There is a need to refactor the dashboard navigation so that it is
a client component and leave the layout as a server component.

[Changes]
- Add DashboardTabs component
- Move dashboard navigation elements to DashboardTabs

CU: https://app.clickup.com/t/86cvpzb2e
